requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-extensions==3.2.1
black==23.3.0
